requirements.txt
gitpython
tabulate
jira
openpyxl
pyyaml
Jinja2
fossology
spdx-tools==0.8.2
